Some SuperIlc cleanup and improvements (dotnet/coreclr#26690)
1) Remove the parameter -cpaot as it's no longer needed now that
Crossgen2 also resides under Core_Root.

2) Add new parameter --exe to only execute the tests without
compilation. JanV specifically asked for this to speed up his
inner testing loop on Linux.

3) Don't pass /largeversionbubble to the [legacy] Crossgen
compilation of System.Private.CoreLib as it apparently crashes
on Linux and we are no longer interested in fixing that.

Thanks

Tomas

Commit migrated from https://github.com/dotnet/coreclr/commit/6ef00c124829ac3b748326bfbeb490387d8ce3d2

+++ src/coreclr/src/tools/ReadyToRun.SuperIlc/Commands/CompileDirectoryCommand.cs
@@ -20,6 +20,12 @@ namespace ReadyToRun.SuperIlc
                 return 1;
             }
 
+            if (options.CoreRootDirectory == null)
+            {
+                Console.Error.WriteLine("--core-root-directory (--cr) is a required argument.");
+                return 1;
+            }
+
             if (options.OutputDirectory == null)
             {
                 options.OutputDirectory = options.InputDirectory;
@@ -33,7 +39,10 @@ namespace ReadyToRun.SuperIlc
 
             IEnumerable<CompilerRunner> runners = options.CompilerRunners(isFramework: false);
 
-            PathExtensions.DeleteOutputFolders(options.OutputDirectory.FullName, options.CoreRootDirectory.FullName, recursive: false);
+            if (!options.Exe)
+            {
+                PathExtensions.DeleteOutputFolders(options.OutputDirectory.FullName, options.CoreRootDirectory.FullName, recursive: false);
+            }
 
             BuildFolder folder = BuildFolder.FromDirectory(options.InputDirectory.FullName, runners, options.OutputDirectory.FullName, options);
             if (folder == null)
@@ -45,7 +54,7 @@ namespace ReadyToRun.SuperIlc
             bool success = folderSet.Build(runners);
             folderSet.WriteLogs();
 
-            if (!options.NoCleanup)
+            if (!options.NoCleanup && !options.Exe)
             {
                 PathExtensions.DeleteOutputFolders(options.OutputDirectory.FullName, options.CoreRootDirectory.FullName, recursive: false);
             }
